requirements/base.txt
-i https://pypi.org/simple
acme==3.3.0; python_version >= '3.9'
aiofiles==24.1.0; python_version >= '3.8'
aiogram==3.22.0; python_version >= '3.9'
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.15; python_version >= '3.9'
aiohttp-socks==0.8.4
aiosignal==1.4.0; python_version >= '3.9'
amqp==5.3.1; python_version >= '3.6'
annotated-types==0.7.0; python_version >= '3.8'
ansible==11.5.0; python_version >= '3.11'
ansible-core==2.18.11; python_version >= '3.11'
ansible-runner==2.4.2; python_version >= '3.9'
anyio==4.11.0; python_version >= '3.9'
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==25.1.0; python_version >= '3.9'
asgiref==3.8.1; python_version >= '3.8'
attrs==25.4.0; python_version >= '3.9'
billiard==4.2.3; python_version >= '3.7'
celery==5.3.6; python_version >= '3.8'
certbot==3.3.0; python_version >= '3.9'
certifi==2025.11.12; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
channels==4.0.0; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
ciso8601==2.3.3
click==8.3.1; python_version >= '3.10'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1.2
click-repl==0.3.0; python_version >= '3.6'
cloudflare==4.1.0; python_version >= '3.8'
configargparse==1.7.1; python_version >= '3.6'
configobj==5.0.9; python_version >= '3.7'
cron-descriptor==2.0.6; python_version >= '3.9'
cryptography==44.0.3; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
distro==1.9.0; python_version >= '3.6'
django==5.2.8; python_version >= '3.10'
django-admin-extra-buttons==1.6.0; python_version >= '3.11'
django-axes==6.3.1; python_version >= '3.7'
django-celery-beat==2.8.1; python_version >= '3.8'
django-celery-results==2.5.1
django-cleanup==8.0.0
django-cors-headers==4.3.1; python_version >= '3.8'
django-crispy-forms==2.5; python_version >= '3.9'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
django-filter==23.5; python_version >= '3.7'
django-graphql-jwt==0.4.0; python_version >= '3.6'
django-htmx==1.21.0; python_version >= '3.9'
django-json-widget==2.0.4
django-jsonform==2.23.2; python_version >= '3.4'
django-netfields==1.3.2
django-ninja==1.4.5; python_version >= '3.7'
django-polymorphic==3.1.0
django-redis==5.4.0; python_version >= '3.6'
django-render-block==0.10; python_version >= '3.8'
django-simple-history==3.8.0; python_version >= '3.9'
django-solo==2.4.0; python_version >= '3.9'
django-taggit==6.1.0; python_version >= '3.8'
django-timezone-field==7.1; python_version >= '3.8' and python_version < '4.0'
django-vite==3.1.0
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-api-key==3.0.0; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
dnspython==2.7.0; python_version >= '3.9'
drf-spectacular==0.27.2; python_version >= '3.7'
flower==2.0.1; python_version >= '3.7'
frozenlist==1.8.0; python_version >= '3.9'
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7; python_version >= '3.7' and python_version < '4'
graphql-relay==3.2.0; python_version >= '3.6' and python_version < '4'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
humanize==4.12.3; python_version >= '3.9'
idna==3.11; python_version >= '3.8'
inertia-django==1.1.0; python_version >= '3.8' and python_version < '4.0'
inflection==0.5.1; python_version >= '3.5'
influxdb-client[ciso]==1.48.0; python_version >= '3.7'
jalali-core==1.0.0; python_version >= '3.8'
jdatetime==5.2.0; python_version >= '3.9'
jinja2==3.1.6; python_version >= '3.7'
josepy==1.15.0; python_version >= '3.8' and python_version < '4.0'
jsonschema==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
kombu==5.6.0; python_version >= '3.9'
lockfile==0.12.2
magic-filter==1.0.12; python_version >= '3.7'
makefun==1.16.0
markupsafe==3.0.3; python_version >= '3.9'
multidict==6.7.0; python_version >= '3.9'
netaddr==1.3.0; python_version >= '3.7'
numpy==2.3.5; python_version >= '3.11'
packaging==25.0; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
parsedatetime==2.6
pexpect==4.9.0
pillow==10.1.0; python_version >= '3.8'
prometheus-client==0.23.1; python_version >= '3.9'
promise==2.3
prompt-toolkit==3.0.52; python_version >= '3.8'
propcache==0.4.1; python_version >= '3.9'
ptyprocess==0.7.0
pycparser==2.23; python_version >= '3.8'
pydantic==2.11.10; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
pyopenssl==25.1.0; python_version >= '3.7'
pyrfc3339==2.1.0; python_version >= '3.9'
python-crontab==3.3.0
python-daemon==3.1.2; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-socks[asyncio]==2.7.3; python_full_version >= '3.8.0'
pytz==2025.2
pyyaml==6.0.3; python_version >= '3.8'
reactivex==4.1.0; python_version >= '3.8' and python_version < '4.0'
redis==7.1.0; python_version >= '3.10'
referencing==0.37.0; python_version >= '3.10'
requests==2.32.5; python_version >= '3.9'
resolvelib==1.0.1
rpds-py==0.29.0; python_version >= '3.10'
sentry-sdk==2.20.0; python_version >= '3.6'
setuptools==80.9.0; python_version >= '3.9'
simpleeval==1.0.3; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
text-unidecode==1.3
tomli-w==1.2.0; python_version >= '3.9'
tornado==6.5.2; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.14; python_version >= '3.6'
whitenoise==6.6.0; python_version >= '3.8'
wonderwords==2.2.0; python_version >= '3.6'
yarl==1.22.0; python_version >= '3.9'
